Разделил компоненты, добавил типы...
