+++ src/app/app.tsx
@@ -5,8 +5,8 @@ import LoginPassword from '../pages/login-screen/login-screen';
 import FavoritesScreen from '../pages/favorites-screen/favorites-screen';
 import OfferScreen from '../pages/offer-screen/offer-screen';
 import NotFoundScreen from '../pages/not-found-screen/not-found-screen';
-import { AppRoute, AuthorizationStatus } from '../../const';
-import PrivateRoute from '../private-route/private-route';
+import { AppRoute, AuthorizationStatus } from '../const';
+import PrivateRoute from '../components/private-route/private-route';
 
 type MainScreenProps = {
   sum: number;
@@ -19,7 +19,8 @@ export default function App({ sum }: MainScreenProps): JSX.Element {
         <Route path={AppRoute.Root} element={<MainScreen sum={sum} />}>
           <Route path={AppRoute.Error404} element={<NotFoundScreen />} />
           <Route path={AppRoute.Login} element={<LoginPassword />} />
-          <Route path={AppRoute.Favorites}
+          <Route
+            path={AppRoute.Favorites}
             element={
               <PrivateRoute authorizationStatus={AuthorizationStatus.NoAuth}>
                 <FavoritesScreen />
