+++ src/components/offers-list/offersList.tsx
@@ -1,0 +1,237 @@
+
+type MainScreenProps = {
+    sum: number;
+  };
+
+export default function OfferList({ sum }: MainScreenProps): JSX.Element {
+    return (
+        <section className="cities__places places">
+        <h2 className="visually-hidden">Places</h2>
+        <b className="places__found">312 places to stay in Amsterdam</b>
+        <form className="places__sorting" action="#" method="get">
+        <span className="places__sorting-caption">Sort by</span>
+        <span className="places__sorting-type" tabIndex={sum}>
+            Popular
+            <svg className="places__sorting-arrow" width="7" height="4">
+            <use xlinkHref="#icon-arrow-select"></use>
+            </svg>
+        </span>
+        <ul className="places__options places__options--custom places__options--opened">
+            <li className="places__option places__option--active" tabIndex={sum}> Popular </li>
+            <li className="places__option" tabIndex={sum}> Price: low to high </li>
+            <li className="places__option" tabIndex={sum}> Price: high to low </li>
+            <li className="places__option" tabIndex={sum}> Top rated first </li>
+        </ul>
+        </form>
+        <div className="cities__places-list places__list tabs__content">
+        <article className="cities__card place-card">
+            <div className="place-card__mark">
+            <span>Premium</span>
+            </div>
+            <div className="cities__image-wrapper place-card__image-wrapper">
+            <a href="#">
+                <img
+                className="place-card__image"
+                src="img/apartment-01.jpg"
+                width="260"
+                height="200"
+                alt="Place image"
+                />
+            </a>
+            </div>
+            <div className="place-card__info">
+            <div className="place-card__price-wrapper">
+                <div className="place-card__price">
+                <b className="place-card__price-value">&euro;120</b>
+                <span className="place-card__price-text">
+                    &#47;&nbsp;night
+                </span>
+                </div>
+                <button className="place-card__bookmark-button button" type="button">
+                <svg className="place-card__bookmark-icon" width="18" height="19">
+                    <use xlinkHref="#icon-bookmark"></use>
+                </svg>
+                <span className="visually-hidden">To bookmarks</span>
+                </button>
+            </div>
+            <div className="place-card__rating rating">
+                <div className="place-card__stars rating__stars">
+                <span style={{ width: '80%' }}></span>
+                <span className="visually-hidden">Rating</span>
+                </div>
+            </div>
+            <h2 className="place-card__name">
+                <a href="#">
+                Beautiful &amp; luxurious apartment at great location
+                </a>
+            </h2>
+            <p className="place-card__type">Apartment</p>
+            </div>
+        </article>
+
+        <article className="cities__card place-card">
+            <div className="cities__image-wrapper place-card__image-wrapper">
+            <a href="#">
+                <img
+                className="place-card__image"
+                src="img/room.jpg"
+                width="260"
+                height="200"
+                alt="Place image"
+                />
+            </a>
+            </div>
+            <div className="place-card__info">
+            <div className="place-card__price-wrapper">
+                <div className="place-card__price">
+                <b className="place-card__price-value">&euro;80</b>
+                <span className="place-card__price-text">
+                    &#47;&nbsp;night
+                </span>
+                </div>
+                <button className="place-card__bookmark-button place-card__bookmark-button--active button" type="button">
+                <svg className="place-card__bookmark-icon" width="18" height="19">
+                    <use xlinkHref="#icon-bookmark"></use>
+                </svg>
+                <span className="visually-hidden">In bookmarks</span>
+                </button>
+            </div>
+            <div className="place-card__rating rating">
+                <div className="place-card__stars rating__stars">
+                <span style={{ width: '80%' }}></span>
+                <span className="visually-hidden">Rating</span>
+                </div>
+            </div>
+            <h2 className="place-card__name">
+                <a href="#">Wood and stone place</a>
+            </h2>
+            <p className="place-card__type">Room</p>
+            </div>
+        </article>
+
+        <article className="cities__card place-card">
+            <div className="cities__image-wrapper place-card__image-wrapper">
+            <a href="#">
+                <img
+                className="place-card__image"
+                src="img/apartment-02.jpg"
+                width="260"
+                height="200"
+                alt="Place image"
+                />
+            </a>
+            </div>
+            <div className="place-card__info">
+            <div className="place-card__price-wrapper">
+                <div className="place-card__price">
+                <b className="place-card__price-value">&euro;132</b>
+                <span className="place-card__price-text">
+                    &#47;&nbsp;night
+                </span>
+                </div>
+                <button className="place-card__bookmark-button button" type="button">
+                <svg className="place-card__bookmark-icon" width="18" height="19">
+                    <use xlinkHref="#icon-bookmark"></use>
+                </svg>
+                <span className="visually-hidden">To bookmarks</span>
+                </button>
+            </div>
+            <div className="place-card__rating rating">
+                <div className="place-card__stars rating__stars">
+                <span style={{ width: '80%' }}></span>
+                <span className="visually-hidden">Rating</span>
+                </div>
+            </div>
+            <h2 className="place-card__name">
+                <a href="#">Canal View Prinsengracht</a>
+            </h2>
+            <p className="place-card__type">Apartment</p>
+            </div>
+        </article>
+
+        <article className="cities__card place-card">
+            <div className="place-card__mark">
+            <span>Premium</span>
+            </div>
+            <div className="cities__image-wrapper place-card__image-wrapper">
+            <a href="#">
+                <img
+                className="place-card__image"
+                src="img/apartment-03.jpg"
+                width="260"
+                height="200"
+                alt="Place image"
+                />
+            </a>
+            </div>
+            <div className="place-card__info">
+            <div className="place-card__price-wrapper">
+                <div className="place-card__price">
+                <b className="place-card__price-value">&euro;180</b>
+                <span className="place-card__price-text">
+                    &#47;&nbsp;night
+                </span>
+                </div>
+                <button className="place-card__bookmark-button button" type="button">
+                <svg className="place-card__bookmark-icon" width="18" height="19">
+                    <use xlinkHref="#icon-bookmark"></use>
+                </svg>
+                <span className="visually-hidden">To bookmarks</span>
+                </button>
+            </div>
+            <div className="place-card__rating rating">
+                <div className="place-card__stars rating__stars">
+                <span style={{ width: '100%' }}></span>
+                <span className="visually-hidden">Rating</span>
+                </div>
+            </div>
+            <h2 className="place-card__name">
+                <a href="#">Nice, cozy, warm big bed apartment</a>
+            </h2>
+            <p className="place-card__type">Apartment</p>
+            </div>
+        </article>
+
+        <article className="cities__card place-card">
+            <div className="cities__image-wrapper place-card__image-wrapper">
+            <a href="#">
+                <img
+                className="place-card__image"
+                src="img/room.jpg"
+                width="260"
+                height="200"
+                alt="Place image"
+                />
+            </a>
+            </div>
+            <div className="place-card__info">
+            <div className="place-card__price-wrapper">
+                <div className="place-card__price">
+                <b className="place-card__price-value">&euro;80</b>
+                <span className="place-card__price-text">
+                    &#47;&nbsp;night
+                </span>
+                </div>
+                <button className="place-card__bookmark-button place-card__bookmark-button--active button" type="button">
+                <svg className="place-card__bookmark-icon" width="18" height="19">
+                    <use xlinkHref="#icon-bookmark"></use>
+                </svg>
+                <span className="visually-hidden">In bookmarks</span>
+                </button>
+            </div>
+            <div className="place-card__rating rating">
+                <div className="place-card__stars rating__stars">
+                <span style={{ width: '80%' }}></span>
+                <span className="visually-hidden">Rating</span>
+                </div>
+            </div>
+            <h2 className="place-card__name">
+                <a href="#">Wood and stone place</a>
+            </h2>
+            <p className="place-card__type">Room</p>
+            </div>
+        </article>
+        </div>
+    </section>
+    )
+}
