--- src/mocks/offers.ts
@@ -1,26 +1,0 @@
-import { Offers } from "../types/types"
-
-
-export function offersCityCard({}:Offers): JSX.Element  {
-    return (
-    [    
-        {
-        "img": [
-            "img/room.jpg",
-            "img/apartment-01.jpg",
-            "img/apartment-02.jpg",
-            "img/apartment-03.jpg"
-        ],
-        "isPremium": Boolean ,
-        "price": [80-220],
-        "title": [
-            "Beautiful &amp; luxurious studio at great location",
-            "Wood and stone place",
-            "Canal View Prinsengracht",
-            "Nice, cozy, warm big bed apartment",
-        ],
-        "type": "apartment",
-        "rating": 4.8,
-        "id": [1-4]
-        }
-    ]
