--- src/types/types.ts
@@ -1,8 +1,0 @@
-export type Offers = {
-    img: string,
-    isPremium: boolean,
-    price: number,
-    title: string,
-    type: string,
-    rating: number
-}
